requirement.txt
kaggle
pyspellchecker
langchain
torch
transformers
faiss-gpu
sentence-transformers
pandas
numpy
accelerate
peft
bitsandbytes
trl
optimum
auto-gptq
streamlit
streamlit-chat
ragas
evaluate